Install rust and enable yjit
This PR installs Rust in the Dockerfile and enables yjit in the images.
The change here is useful for using and testing yjit without being
required to manually configure Ruby. One usecase is to add a build to
rails/rails to test our code against yjit.

I tested this locally and was able to build the docker image and
verified it built correctly and I could use yjit by running the
following command:

```
$ RUBY_YJIT_ENABLE=1 ruby -e 'p RubyVM::YJIT.enabled?'
=> true
```